requirements.txt
influxdb
pycryptodome